requirements.txt
praw
mysql-connector-python
python-dotenv
pandas
